Create a test server that can be used for the review pagination test.
